+++ src/Equinor.ProCoSys.Preservation.WebApi/Program.cs
@@ -1,7 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Reflection;
-using Azure.Core;
-using Azure.Identity;
 using Equinor.ProCoSys.Auth;
 using Equinor.ProCoSys.Completion.WebApi.DIModules;
 using Equinor.ProCoSys.Preservation.Command;
@@ -9,7 +7,6 @@ using Equinor.ProCoSys.Preservation.Query;
 using Equinor.ProCoSys.Preservation.WebApi.DiModules;
 using Equinor.ProCoSys.Preservation.WebApi.DIModules;
 using Equinor.ProCoSys.Preservation.WebApi.Middleware;
-using Equinor.ProCoSys.Preservation.WebApi.Misc;
 using FluentValidation;
 using FluentValidation.AspNetCore;
 using Microsoft.AspNetCore.Builder;
@@ -17,32 +14,10 @@ using Microsoft.AspNetCore.Http.Features;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Identity.Web;
 
 var builder = WebApplication.CreateBuilder(args);
 
-var devOnLocalhost = builder.Configuration.IsDevOnLocalhost();
-
-// ChainedTokenCredential iterates through each credential passed to it in order, when running locally
-// DefaultAzureCredential will probably fail locally, so if an instance of Azure Cli is logged in, those credentials will be used
-// If those credentials fail, the next credentials will be those of the current user logged into the local Visual Studio Instance
-// which is also the most likely case
-TokenCredential credential = devOnLocalhost switch
-{
-    true
-        => new ChainedTokenCredential(
-            new AzureCliCredential(),
-            new VisualStudioCredential(),
-            new DefaultAzureCredential()
-        ),
-    false => new DefaultAzureCredential()
-};
-
-builder.Services.AddSingleton(credential);
-
-builder.Services.AddMicrosoftIdentityWebApiAuthentication(builder.Configuration)
-    .EnableTokenAcquisitionToCallDownstreamApi()
-    .AddInMemoryTokenCaches();
+builder.ConfigureAuthentication();
 
 builder.ConfigureHttp();
 
@@ -104,7 +79,7 @@ app.UseHttpsRedirection();
 
 app.UseRouting();
 
-// order of adding middelwares are crucial. Some depend that other has been run in advance
+// order of adding middlewares are crucial. Some depend on that other has been run in advance
 app.UseCurrentPlant();
 app.UseAuthentication();
 app.UseCurrentUser();
@@ -119,4 +94,4 @@ app.MapControllers();
 
 app.Run();
 
-public partial class Program;
+public abstract partial class Program;
